linting and organizing
