+++ src/js/TangramPlay.js
@@ -16,7 +16,8 @@ import LocalStorage from 'app/addons/LocalStorage';
 
 // Import Utils
 import xhr from 'xhr';
-import { StopWatch, subscribeMixin, debounce, createObjectURL } from 'app/core/common';
+import { subscribeMixin } from 'app/tools/mixin';
+import { StopWatch, debounce, createObjectURL } from 'app/tools/common';
 import { selectLines, isStrEmpty } from 'app/core/codemirror/tools';
 import { getNodes, parseYamlString } from 'app/core/codemirror/yaml-tangram';
