+++ src/js/addons/pickers/ColorPicker.js
@@ -5,27 +5,25 @@ Author: Lou Huang (@saikofish)
 
 'use strict';
 
-import Picker from './Picker'
-import Color from './Color';
-import { getDevicePixelRatio, addEvent, removeEvent } from './common'
+import Picker from './Picker';
+import Color from './types/Color';
+import { addEvent, removeEvent } from './Picker';
+import { getDevicePixelRatio } from 'app/tools/common';
 
 // Some common use variables
 let currentTarget;
 let currentTargetHeight = 0;
 let domCache;
 
-// Default modal appearance values
-const MODAL_VIEWPORT_EDGE_BUFFER = 20; // buffer zone at the viewport edge where a modal should not be presented
-
 export default class ColorPicker extends Picker {
-    constructor (color = 'vec3(1.0,0.0,0.0)', properties) {
+    constructor (color = 'vec3(1.0,0.0,0.0)', properties = {}) {
         super('colorpicker-', properties);
-        
-        this.width = 250;  // in pixels
+
+        this.width = 250; // in pixels
         this.height = 250; // in pixels
 
-        this.disc = {width: 200, height: 200 };
-        this.barlum = {width: 25, height: 200};
+        this.disc = { width: 200, height: 200 };
+        this.barlum = { width: 25, height: 200 };
 
         this.setValue(color);
         this.init();
@@ -45,7 +43,7 @@ export default class ColorPicker extends Picker {
             let barcursors = document.createElement('div');
             let leftcursor = document.createElement('div');
             let rightcursor = document.createElement('div');
-            
+
             modal.className = this.CSS_PREFIX + 'modal picker-modal';
             patch.className = this.CSS_PREFIX + 'patch';
             map.className = this.CSS_PREFIX + 'hsv-map';
@@ -58,17 +56,13 @@ export default class ColorPicker extends Picker {
             barcursors.className = this.CSS_PREFIX + 'bar-cursors';
             leftcursor.className = this.CSS_PREFIX + 'bar-cursor-left';
             rightcursor.className = this.CSS_PREFIX + 'bar-cursor-right';
-            
-            // disc.width = this.disc.width;
-            // disc.height = this.disc.height;
-            // barlum.width = this.barlum.width;
-            // barlum.height = this.barlum.height;
+
             map.id = 'cp-map';
             barcursors.id = 'cp-bar';
 
             modal.appendChild(patch);
             modal.appendChild(map);
-            
+
             map.appendChild(disc);
             map.appendChild(cover);
             map.appendChild(cursor);
@@ -96,7 +90,7 @@ export default class ColorPicker extends Picker {
         this.dom.hsvBarCursors = this.dom.hsvMap.children[6];
         this.dom.hsvLeftCursor = this.dom.hsvBarCursors.children[0];
         this.dom.hsvRightCursor = this.dom.hsvBarCursors.children[1];
-        
+
         this.dom.colorDisc = this.el.querySelector('.colorpicker-disc');
         this.dom.luminanceBar = this.el.querySelector('.colorpicker-bar-luminance');
 
@@ -105,8 +99,8 @@ export default class ColorPicker extends Picker {
             lbutton.innerHTML = '+';
             lbutton.className = this.CSS_PREFIX + 'link-button';
             this.el.appendChild(lbutton);
-            // this.dom.lbutton = this.el.querySelector('.colorpicker-link-button');
-            lbutton.addEventListener("click", () => {
+
+            lbutton.addEventListener('click', () => {
                 this.trigger('link_button', this.value);
                 if (typeof this.link_button === 'function') {
                     this.link_button(this.value);
@@ -116,7 +110,7 @@ export default class ColorPicker extends Picker {
         }
     }
 
-    draw () {     
+    draw () {
         //  Render color patch
         let patch = this.el.querySelector('.colorpicker-patch');
         patch.style.backgroundColor = this.value.getString('rgb');
@@ -129,7 +123,7 @@ export default class ColorPicker extends Picker {
         let y = Math.sin(pi2 - color.hsv.h * pi2);
         let r = color.hsv.s * (colorDiscRadius - 5);
 
-        this.dom.hsvMapCover.style.opacity = 1 - color.hsv.v/255;
+        this.dom.hsvMapCover.style.opacity = 1 - color.hsv.v / 255;
         // this is the faster version...
         this.dom.hsvBarWhiteLayer.style.opacity = 1 - color.hsv.s;
         this.dom.hsvBarBGLayer.style.backgroundColor = 'rgb(' +
@@ -150,13 +144,13 @@ export default class ColorPicker extends Picker {
         }
 
         if (this.dom.hsvLeftCursor) {
-            this.dom.hsvLeftCursor.style.top = this.dom.hsvRightCursor.style.top = ((1 - color.hsv.v/255) * colorDiscRadius * 2) + 'px';
+            this.dom.hsvLeftCursor.style.top = this.dom.hsvRightCursor.style.top = ((1 - color.hsv.v / 255) * colorDiscRadius * 2) + 'px';
         }
     }
 
     presentModal (x, y) {
-        super.presentModal(x,y);
-    
+        super.presentModal(x, y);
+
         // // Listen for interaction on the HSV map
         this.onHsvDownHandler = addEvent(this.dom.hsvMap, 'mousedown', this.onHsvDown, this);
 
@@ -164,18 +158,18 @@ export default class ColorPicker extends Picker {
 
         if (colorDisc.getContext) {
             // HSV color wheel with white center
-            let diskContex = colorDisc.getContext('2d');
-            let ratio = getDevicePixelRatio(diskContex);
+            let diskContext = colorDisc.getContext('2d');
+            let ratio = getDevicePixelRatio(diskContext);
             let width = this.disc.width / ratio;
             let height = this.disc.height / ratio;
             this.dom.colorDisc.width = width * ratio;
-            this.dom.colorDisc.height = height* ratio;
-            diskContex.scale(ratio, ratio);
+            this.dom.colorDisc.height = height * ratio;
+            diskContext.scale(ratio, ratio);
 
             drawDisk(
-                diskContex,
-                [width/2, height/2],
-                [width/2 - 1, height/2 - 1],
+                diskContext,
+                [width / 2, height / 2],
+                [width / 2 - 1, height / 2 - 1],
                 360,
                 function (ctx, angle) {
                     let gradient = ctx.createRadialGradient(1, 1, 1, 1, 1, 0);
@@ -189,11 +183,11 @@ export default class ColorPicker extends Picker {
 
             // gray border
             drawCircle(
-                diskContex,
-                [width/2, height/2],
-                [width/2, height/2],
+                diskContext,
+                [width / 2, height / 2],
+                [width / 2, height / 2],
                 '#303030',
-                2/ratio
+                2 / ratio
             );
 
             // draw the luminanceBar bar
@@ -201,7 +195,7 @@ export default class ColorPicker extends Picker {
             this.dom.luminanceBar.width = this.barlum.width ;
             this.dom.luminanceBar.height = this.barlum.height * ratio;
             ctx.scale(ratio, ratio);
-            let gradient = ctx.createLinearGradient(0, 0, 0, this.barlum.height/ratio);
+            let gradient = ctx.createLinearGradient(0, 0, 0, this.barlum.height / ratio);
 
             gradient.addColorStop(0, 'transparent');
             gradient.addColorStop(1, 'black');
@@ -209,7 +203,6 @@ export default class ColorPicker extends Picker {
             ctx.fillStyle = gradient;
             ctx.fillRect(0, 0, 30, 200);
         }
-        
         this.draw();
     }
 
@@ -254,13 +247,13 @@ export default class ColorPicker extends Picker {
             r = currentTargetHeight / 2,
             x = event.offsetX - r,
             y = event.offsetY - r,
-            h = (360 - ((Math.atan2(y, x) * 180 / Math.PI) + (y < 0 ? 360 : 0)))/360,
+            h = (360 - ((Math.atan2(y, x) * 180 / Math.PI) + (y < 0 ? 360 : 0))) / 360,
             s = (Math.sqrt((x * x) + (y * y)) / r);
             this.value.set({ h, s }, 'hsv');
         }
         else if (event.target === this.dom.hsvBarCursors && currentTarget === this.dom.hsvBarCursors) { // the luminanceBar
             let v = (currentTargetHeight - (event.offsetY)) / currentTargetHeight;
-            v = Math.max(0, Math.min(1, v))*255;
+            v = Math.max(0, Math.min(1, v)) * 255;
             this.value.set({ v: v }, 'hsv');
         }
