+++ src/js/addons/pickers/Vec3Picker.js
@@ -1,11 +1,9 @@
 'use strict';
 
-import Picker from './Picker'
-import Vector from './Vector';
-import Matrix from './Matrix';
-import { getDevicePixelRatio, addEvent, removeEvent } from './common'
-
-let domCache;
+import Picker from './Picker';
+import Vector from './types/Vector';
+import Matrix from './types/Matrix';
+import { addEvent, removeEvent } from './Picker';
 
 export default class Vec3Picker extends Picker {
     constructor (dir, properties) {
@@ -13,34 +11,32 @@ export default class Vec3Picker extends Picker {
 
         this.width = this.width || 200;
         this.height = this.width || 200;
-        this.fnColor = this.fnColor || 'rgb(230, 230, 230)';
-        this.dimColor = this.dimColor || 'rgb(100, 100, 100)';
-        this.dragScale = 50;
+        this.scale = 50;
 
-        this.setValue(dir || [0,0,1]);
+        this.setValue(dir || [0, 0, 1]);
         this.create();
-        
+
         this.camera = new Matrix();
         this.shapes = [];
         this.center = [0, 0, 0];
 
         this.shapes.push({
             edgeColour: this.dimColor,
-            nodes: [[this.width/2-50, this.height/2, 100], [this.width/2+50,  this.height/2,  100],
-                    [this.width/2, this.height/2-50, 100], [this.width/2, this.height/2+50, 100],
-                    [this.width/2, this.height/2, 50], [this.width/2, this.height/2, 150] ],
+            nodes: [[this.width / 2 - 50, this.height / 2, 100], [this.width / 2 + 50, this.height / 2, 100],
+                    [this.width / 2, this.height / 2 - 50, 100], [this.width / 2, this.height / 2 + 50, 100],
+                    [this.width / 2, this.height / 2, 50], [this.width / 2, this.height / 2, 150] ],
             edges: [[0,1], [2,3], [4,5]]
         });
 
         this.shapes.push({
             textColour: this.fnColor,
-            nodes: [[this.width/2+68, this.height/2, 100], [this.width/2-68, this.height/2, 100], 
-                    [this.width/2, this.height/2+68, 100], [this.width/2, this.height/2-68, 100], 
-                    [this.width/2, this.height/2, 168], [this.width/2, this.height/2, 32] ],
-            text: ["x", "-x", "y", "-y", "z", "-z"]
+            nodes: [[this.width / 2 + 68, this.height / 2, 100], [this.width / 2 - 68, this.height / 2, 100],
+                    [this.width / 2, this.height / 2 + 68, 100], [this.width / 2, this.height / 2 - 68, 100],
+                    [this.width / 2, this.height / 2, 168], [this.width / 2, this.height / 2, 32] ],
+            text: ['x', '-x', 'y', '-y', 'z', '-z']
         });
 
-        this.setCenter(this.width/2,this.height/2,100);
+        this.setCenter(this.width / 2, this.height / 2, 100);
 
         // Mouse events
         this.dragOffset = [0, 0];
@@ -50,7 +46,7 @@ export default class Vec3Picker extends Picker {
     setCenter (x, y, z) {
         for (let s in this.shapes) {
             let shape = this.shapes[s];
-            
+
             for (let n in shape.nodes) {
                 shape.nodes[n][0] -= x;
                 shape.nodes[n][1] -= y;
@@ -64,10 +60,10 @@ export default class Vec3Picker extends Picker {
         let A = this.camera.getMult(node);
         A.add(this.center);
         return [A.x, this.height - A.y];
-    };
+    }
 
-	draw () {    
-        this.ctx.clearRect(0,0,this.width,this.height);
+    draw () {
+        this.ctx.clearRect(0, 0, this.width, this.height);
 
         for (let s in this.shapes) {
             let shape = this.shapes[s];
@@ -89,11 +85,11 @@ export default class Vec3Picker extends Picker {
         });
 
         this.drawShapeNodes({
-            nodeColour: this.overPoint? '#28A86B' : this.fnColor,
-            nodeRadius: this.overPoint? 4 : 2,
+            nodeColour: this.overPoint ? this.selColor : this.fnColor,
+            nodeRadius: this.overPoint ? 4 : 2,
             nodes: [this.point]
         });
-	}
+    }
 
     drawShapeEdges (shape) {
         let nodes = shape.nodes;
@@ -121,7 +117,7 @@ export default class Vec3Picker extends Picker {
         }
     }
 
-    drawShapeText (shape) {        
+    drawShapeText (shape) {
         this.ctx.fillStyle = shape.textColour;
         for (let n in shape.nodes) {
             let coord = this.viewFromCamera(shape.nodes[n]);
@@ -152,16 +148,16 @@ export default class Vec3Picker extends Picker {
 
         if (this.overPoint) {
             let invM = this.camera.getInv();
-            let vel = invM.getMult([dx,-dy,-0.00001]);
+            let vel = invM.getMult([dx, -dy, -0.00001]);
+            vel.mult(2);
             this.value.add(vel);
-            // this.value.normalize();
-            this.point = [this.value.x*this.dragScale, this.value.y*this.dragScale, this.value.z*this.dragScale];
+            this.point = [this.value.x * this.scale, this.value.y * this.scale, this.value.z * this.scale];
         }
         else {
             this.camera.rotateX(dy);
-            this.camera.rotateY(dx); 
+            this.camera.rotateY(dx);
         }
-        
+
         this.dragOffset = [x, y];
 
         // fire 'changed'
@@ -169,20 +165,49 @@ export default class Vec3Picker extends Picker {
     }
 
     onDbClick (event) {
+        let mouse = new Vector([event.offsetX, event.offsetY]);
+        let axis = {
+            x: [68, 0, 0],
+            neg_x: [-68, 0, 0],
+            y: [0, 68, 100],
+            neg_y: [0, -68, 0]
+        };
+        let selected = '';
+        for (let i in axis) {
+            let pos = new Vector(this.viewFromCamera(axis[i]));
+            let diff = pos.getSub(mouse);
+            if (diff.getLength() < 10) {
+                selected = i;
+                break;
+            }
+        }
         this.camera = new Matrix();
+
+        if (selected === 'x') {
+            this.camera.rotateY(-1.57079632679);
+        }
+        else if (selected === 'neg_x') {
+            this.camera.rotateY(1.57079632679);
+        }
+        else if (selected === 'y') {
+            this.camera.rotateX(-1.57079632679);
+        }
+        else if (selected === 'neg_y') {
+            this.camera.rotateX(1.57079632679);
+        }
+
         this.draw();
     }
 
     destroyEvents () {
-        super.destroyEvents()
+        super.destroyEvents();
         removeEvent(this.el, 'dblclick', this.onDbClick);
         this.onMouseMoveHandler = null;
     }
 
     setValue (dir) {
         this.value = new Vector(dir);
-        // this.value.normalize();
-        this.point = [this.value.x*this.dragScale, this.value.y*this.dragScale, this.value.z*this.dragScale];
+        this.point = [this.value.x * this.scale, this.value.y * this.scale, this.value.z * this.scale];
     }
 }
