+++ src/js/addons/ui/Modal.Examples.js
@@ -3,7 +3,7 @@
 import TangramPlay, { container } from 'app/TangramPlay';
 import Modal from 'app/addons/ui/Modal';
 import EditorIO from 'app/addons/ui/EditorIO';
-import { httpGet } from 'app/core/common';
+import { httpGet } from 'app/tools/common';
 
 let examplesEl;
